+++ ledger/api/__init__.py
@@ -1,12 +1,13 @@
+import logging
+
 from ninja import NinjaAPI
 from ninja.security import django_auth
 
 from django.conf import settings
 
 from ledger.api import ledger
-from ledger.hooks import get_extension_logger
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 api = NinjaAPI(
     title="Geuthur API",