+++ ledger/managers/characterjournal_manager.py
@@ -1,3 +1,4 @@
+import logging
 from collections import defaultdict
 
 from django.db import models
@@ -7,11 +8,10 @@ from django.utils import timezone
 
 from allianceauth.eveonline.models import EveCharacter
 
-from ledger.hooks import get_extension_logger
 from ledger.managers.manager_helper import _annotations_information
 from ledger.view_helpers.core import events_filter
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 # PvE - Income
 BOUNTY_PRIZES = ["bounty_prizes"]
 ESS_TRANSFER = ["ess_escrow_transfer"]