+++ ledger/task_helpers/core_helpers.py
@@ -2,11 +2,11 @@
 Core Helpers
 """
 
+import logging
+
 from esi.models import Token
 
-from ledger.hooks import get_extension_logger
-
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 def get_token(character_id: int, scopes: list) -> Token: