+++ ledger/views/alliance/alliance_ledger.py
@@ -1,5 +1,6 @@
 """PvE Views"""
 
+import logging
 from datetime import datetime
 
 # Django
@@ -7,12 +8,10 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
 from django.shortcuts import redirect, render
 
-from ledger.hooks import get_extension_logger
-
 # Ledger
 from ledger.view_helpers.core import add_info_to_context
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 @login_required